requirements.txt
Django>=1.11
selenium
factory_boy